requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
cryptography==44.0.2
django==3.2
djongo==1.3.6
sqlparse==0.2.4
djangorestframework==3.13.1
django-cors-headers==3.13.0
djangorestframework-simplejwt==4.8.0
dnspython==2.6.1
dotenv==0.9.9
mongoengine==0.28.2
pycparser==2.22
pymongo==3.12.3
python-dotenv==1.1.0
pytz==2024.1
typing_extensions==4.12.2
tzdata==2024.1